pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "goats"
description = "Gemini Observation and Analysis of Targets System."
readme = "README.md"
authors = [
    { name = "Dan Avner", email = "dan.avner@noirlab.edu" },
    { name = "Monika Soraisam", email = "monika.soraisam@noirlab.edu" },
]
maintainers = [{ name = "GOATS Team", email = "goats@noirlab.edu" }]
classifiers = [
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Astronomy",
]
requires-python = ">=3.12"
dependencies = [
    "astropy>=6.0,<7",
    "astroquery>=0.4.10",
    "channels-redis>=4.2.1,<5",
    "channels[daphne]>=4.0,<5",
    "click>=8.2.1,<9",
    "django>=4.2.23,<5",
    "django-cors-headers>=4.7.0,<5",
    "django-dramatiq>=0.13.0",
    "djangorestframework>=3.16.0,<4",
    "dramatiq-abort>=1.2.1",
    "dramatiq[redis,watch]>=1.18.0,<2",
    "gpp-client>=25.6.0",
    "marshmallow>=3.26.1,<4",
    "marshmallow-jsonapi>=0.24.0",
    "numpydoc>=1.9.0,<2",
    "tomtoolkit>=2.24.5",
]
version = "25.6.0"

[project.urls]
Homepage = "https://github.com/gemini-hlsw/goats"
Documentation = "https://goats.readthedocs.io/en/latest/"
Repository = "https://github.com/gemini-hlsw/goats"
Issues = "https://github.com/gemini-hlsw/goats/issues"
Changelog = "https://github.com/gemini-hlsw/goats/blob/main/CHANGES.md"

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "nox>=2025.5.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
    "pytest-remotedata>=0.4.1",
    "pytest-xdist>=3.7.0",
    "ruff>=0.11.13",
    "towncrier>=24.8.0",
]
docs = [
    "furo>=2024.8.6",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinxcontrib-video>=0.4.1",
]
github = [
    "dragons",
    "fits-storage",
]
notebook = ["ipykernel>=6.29.5"]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
goats = "goats_cli:cli"

[tool.setuptools.package-data]
"goats_tom.templates" = ["**/*.html"]
"goats_tom.static" = ["**/*.*"]
"goats_setup.templates" = ["**/*.tmpl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "goats_tom.tests.settings"
norecursedirs = "tests/unit/goats_tom/ocs"
asyncio_default_fixture_loop_scope = "session"
addopts = "-r A -v -n auto"

[tool.towncrier]
version = ""
package_dir = "goats"
filename = "docs/readthedocs/source/project_details/changelog.rst"
underlines = "=-^"
directory = "docs/changes"
start_string = ".. towncrier release notes start"
title_format = "Version {version} ({project_date})"
issue_format = "`PR #{issue} <https://github.com/gemini-hlsw/goats/pull/{issue}>`_"
[[tool.towncrier.type]]
directory = "new"
name = "New Features"
showcontent = true
[[tool.towncrier.type]]
directory = "change"
name = "Changes"
showcontent = true
[[tool.towncrier.type]]
directory = "bugfix"
name = "Bug Fixes"
showcontent = true
[[tool.towncrier.type]]
directory = "perf"
name = "Enhancements"
showcontent = true
[[tool.towncrier.type]]
directory = "other"
name = "Other"
showcontent = true
[[tool.towncrier.type]]
directory = "doc"
name = "Documentation"
showcontent = true

[tool.ruff]
extend-exclude = [
    "src/goats_tom/migrations",
    "src/goats_tom/facilities/gemini.py",
    "src/goats_tom/tests/settings.py",
    "src/goats_tom/tests/urls.py",
    "docs/",
    "tests/"
]
target-version = "py312"

[tool.ruff.lint]
select = ['E', 'F', 'W', 'A', 'PLC', 'PLE', 'PLW', 'I']
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.uv.sources]
dragons = { git = "https://github.com/GeminiDRSoftware/DRAGONS", tag = "v4.0.0" }
fits-storage = { git = "https://github.com/GeminiDRSoftware/FitsStorage", tag = "v3.4.0" }
